requirements.txt
Django==2.1.4
django-ipware==2.1.0
pytz==2018.7
